+++ src/Helpers/Customer.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Webkul\Faker\Helpers;
+
+use Webkul\Customer\Models\Customer as CustomerModel;
+
+class Customer
+{
+    /**
+     * Create a records
+     *
+     * @param  integer  $count
+     * @return void
+     */
+    public function create($count)
+    {
+        CustomerModel::factory()
+            ->count($count)
+            ->create();
+    }
+}