+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 #### This changelog consists the bug & security fixes and new features being included in the releases listed below.
 
+## **v2.0.0-ALPHA1(14-Oct-2022)** - *Release*
+
+* [Feature] Now creates Simple, Virtual, Downloadable and Configurable products.
+
+* [Feature] Added support to add 10 records in a single batch using queue system
+
+* [Enhancement] Refactored entire code base
+
+
 ## **v1.3.3(24-Jan-2022)** - *Release*
 
-* [compatibility] Compatible with bagisto v1.3.3.
+* [compatibility] Compatible with Bagisto v1.3.3.
 
 * [compatibility] Seed fake data using Laravel 8 Modal Factory.